requirements.txt
accelerate>=1.4.0
fastapi>=0.115.0
Jinja2>=3.1.2
MarkupSafe>=2.1.1
numpy>=1.26.0
opencv-python>=4.9.0
pillow>=10.0.0
pydantic>=2.8.0
pyrealsense2
python-multipart>=0.0.9
regex>=2024.11.6
requests>=2.31.0
sentencepiece>=0.2.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.49.0
ultralytics>=8.3.0
uvicorn>=0.30.0
